requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
openai
python-whois
motor
python-multipart
pyzbar
opencv-python-headless
requests
aiohttp
python-dotenv
joblib
pydantic
tldextract
validators
